@@ -1,25 +1,36 @@
 {
-  "name": "GoldPriceTracking",
-  "version": "1.0.1",
+  "name": "gold-price-tracking",
+  "version": "1.2.0",
   "private": true,
   "devDependencies": {
-    "react-scripts": "1.0.14"
+    "react-scripts": "^3.0.1"
   },
   "dependencies": {
-    "axios": "^0.16.2",
-    "element-react": "^1.4.5",
-    "element-theme-default": "^1.4.13",
-    "eslint": "^4.9.0",
-    "moment-timezone": "^0.5.13",
-    "react": "^16.0.0",
-    "react-dom": "^16.0.0",
-    "react-vis": "^1.7.9",
-    "recharts": "^1.0.0-apha.5"
+    "@material-ui/core": "^4.1.1",
+    "@material-ui/icons": "^4.2.0",
+    "axios": "^0.19.0",
+    "eslint": "^5.16.0",
+    "moment-timezone": "^0.5.25",
+    "react": "^16.8.6",
+    "react-dom": "^16.8.6",
+    "recharts": "^1.6.2"
   },
   "scripts": {
     "start": "react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test --env=jsdom",
     "eject": "react-scripts eject"
+  },
+  "browserslist": {
+    "production": [
+      ">0.2%",
+      "not dead",
+      "not op_mini all"
+    ],
+    "development": [
+      "last 1 chrome version",
+      "last 1 firefox version",
+      "last 1 safari version"
+    ]
   }
 }
