@@ -1,68 +1,104 @@
-import React, { Component } from 'react';
-import { Table } from 'element-react';
-import Filter from './Filter';
+import React from 'react';
+import { makeStyles, withStyles } from '@material-ui/core/styles';
+import Table from '@material-ui/core/Table';
+import TableBody from '@material-ui/core/TableBody';
+import TableCell from '@material-ui/core/TableCell';
+import TableHead from '@material-ui/core/TableHead';
+import TableRow from '@material-ui/core/TableRow';
+import TablePagination from '@material-ui/core/TablePagination';
+import Paper from '@material-ui/core/Paper';
 import moment from 'moment-timezone';
+import Filter from './Filter';
+import './Table.css';
 
-class GoldTable extends Component {
-    constructor(props) {
-        super(props);
-        this.state = {
-            coloums: [
-                {
-                    label: 'Date',
-                    prop: 'created_at',
-                    width: window.innerWidth * 0.4
-                },
-                {
-                    label: 'Buy(รับซื้อ)',
-                    prop: 'buy',
-                    width: window.innerWidth * 0.2
-                },
-                {
-                    label: 'Sell',
-                    prop: 'sell',
-                    width: window.innerWidth * 0.2
-                }
-            ],
-            dataForTable: null,
-        };
+function GoldTable(props) {
+    const [page, setPage] = React.useState(0);
+    const [rowsPerPage, setRowsPerPage] = React.useState(10);
+    const classes = makeStyles(theme => ({
+        root: {
+            width: '100%',
+            marginTop: theme.spacing(1),
+            overflowX: 'auto',
+        },
+        table: {
+            minWidth: 300,
+        },
+    }))();
+    const rows = mapDataForTable(props.prices);
+    const StyledTableCell = withStyles(theme => ({
+        head: {
+            backgroundColor: theme.palette.common.black,
+            color: theme.palette.common.white,
+        },
+        body: {
+            fontSize: 14
+        }
+    }))(TableCell);
+
+    function mapDataForTable(goldPrices) {
+        if (goldPrices === null || goldPrices.length === 0) {
+            return [];
+        }
+
+        console.log(goldPrices)
+
+        return goldPrices.map(x => Object.create({
+            date: moment(x.created_at).tz('Asia/Bangkok').format('YYYY-MM-DD HH:mm'),
+            buy: x.buy,
+            sell: x.sell
+        })).reverse();
     }
 
-    retrieveGoldPrices = () => {
-        
+    function handleChangePage(event, newPage) {
+        setPage(newPage);
     }
 
-    mapDataForTable = (goldPrices) => {
-        let data = goldPrices.map(x => {
-            let newData = {};
-            newData['buy'] = x.buy;
-            newData['sell'] = x.sell;
-            newData['created_at'] = moment(x.created_at).tz('Asia/Bangkok').format('LLL');
-            return newData;
-        });
-        this.setState({
-            dataForTable: data
-        });
+    function handleChangeRowsPerPage(event) {
+        setRowsPerPage(+event.target.value);
     }
 
-    componentWillMount() {
-       
-    }
-    
-    render() {
-        return (
-            <div>
-                {/* <Filter onChangeDaysBack={this.props.onChangeDaysBack}/> */}
-                <Table
-                    style={{width: '100%', marginLeft: '1%', marginRight: '1%'}}
-                    columns={this.state.columns}
-                    data={this.state.dataForTable}
-                    border={true}
-                    height={window.innerHeight * 0.5}
-                />
+    return (
+        <div className='table'>
+            <div className="filter">
+                <Filter setValue={props.onChangeNumOfRec} />
             </div>
-        );
-    }
+            <Paper className={classes.root}>
+                <Table className={classes.table} size="small">
+                    <TableHead>
+                        <TableRow>
+                            <StyledTableCell align='center'>Date</StyledTableCell>
+                            <StyledTableCell align='center'>Buy(baht)</StyledTableCell>
+                            <StyledTableCell align='center'>Sell(baht)</StyledTableCell>
+                        </TableRow>
+                    </TableHead>
+                    <TableBody>
+                        {rows.slice(page * rowsPerPage, page * rowsPerPage + rowsPerPage).map((row, idx) => (
+                            <TableRow key={idx}>
+                                <TableCell align='center'>{row.date}</TableCell>
+                                <TableCell align='center'>{row.buy}</TableCell>
+                                <TableCell align='center'>{row.sell}</TableCell>
+                            </TableRow>
+                        ))}
+                    </TableBody>
+                </Table>
+                <TablePagination
+                    rowsPerPageOptions={[10, 25, 50]}
+                    component="div"
+                    count={rows.length}
+                    rowsPerPage={rowsPerPage}
+                    page={page}
+                    backIconButtonProps={{
+                        'aria-label': 'Previous Page',
+                    }}
+                    nextIconButtonProps={{
+                        'aria-label': 'Next Page',
+                    }}
+                    onChangePage={handleChangePage}
+                    onChangeRowsPerPage={handleChangeRowsPerPage}
+                />
+            </Paper>
+        </div>
+    );
 }
 
 export default GoldTable;
